Refactoring some functions a bit. Now, functions related to vrep is in utils_vrep.py
